Create get_usermacro.py
Retrieve a macro value for a host.  Hopefully this could be used as an external script like get_usermacro.py[{HOST.HOST}, "{$SOME_MACRO}"] to store a macro value into an item.  The item could be set to update a host inventory field.  

An update interval of 24 hours should be adequate unless the macro value changes frequently.,